Suggestions screen
Fixed problem with some classes being "mysteriously" unassigned when Suggestion screen was used on a solution loaded using automated (not-interactive) solver due to:
	- section Placements of Suggestions screen is used
	- a wrong class get unassigned during assignment of a suggestion

The problem was caused by constraints that returns a set of conflicting constraints in a non-deterministic way (e.g., spread in time constraint with initial allowance over 2 classes per a time slot -- e.g., spread of 21 or more classes over 10 possible times). Undo of a tested placement or assignment of a suggestion was done in a way that did not anticipated a possibility of a different set of conflicting classes to be returned (different than during assignment of the tested placement / suggestion). To overcome this problem, all assignments that are to be changed need to be unassigned first, and then assigned to the new values.


git-svn-id: http://unitime.googlecode.com/svn/trunk@242 cc804148-bb53-fe5c-6b9d-7ba62d325a45