Added support for meetsWith element to offerings xml.
If useMeetsWith="true" on offerings then the meetsWith element is parsed if it is present on a class.

The meetsWith element is used to create a meets with distribution preference on the classes listed as meets with.

This element only works if you are using external unique ids for classes.

If there are other meets with preferences on a class at the time of the xml load they will be removed and replaced by a new meets with that contains only the classes listed as meets with in the xml file.


git-svn-id: http://unitime.googlecode.com/svn/trunk@605 cc804148-bb53-fe5c-6b9d-7ba62d325a45