In some cases the user needs to have more than one class assigned to the same room at the same time.  The database solver did not allow for that situation to occur.  I have added support to allow the user to check a box that says "Do not unassign conflicting class assignments" in the room filter.  Checking this box tells the database solver to ignore room conflicts and instructor conflicts when making the assignment.  This prevents the database solver from unassigning classes with these types of conflicts when the user puts a class into a room.
Also, locations that are marked as ignore room checks no longer cause classes that have conflicting assignments to become unassigned when an additional class is assigned to the location.  The location is still shown with a line through it if there are conflicting assignments so the user knows they are selecting a location that has conflicts.


git-svn-id: http://unitime.googlecode.com/svn/trunk@1419 cc804148-bb53-fe5c-6b9d-7ba62d325a45