Added a set of Enrollment Audit PDF reports.
These reports can be found as a new menu item under the "Student Sectioning" sidebar menu.

There are three new reports.

Enrollments Violating Course Structure - This report lists where students are enrolled in a class that has a parent class but the student is not enrolled in the parent class.

Missing Course Enrollments - This report lists where student are enrolled in a course offering but are not enrolled in a class in each of the subparts within a configuration of the coures offering.

Multiple Course Enrollments - This reports lists where students are enrolled in more than one class within a subpart.  If a student is allowed to enroll in a course offering more than once in a term this student will be listed here as well.


git-svn-id: http://unitime.googlecode.com/svn/trunk@1502 cc804148-bb53-fe5c-6b9d-7ba62d325a45